Обработчик кнопки show more перенесен в presenter
